add to example plugin
