+++ .tree-sitter-lint/tree-sitter-lint-local/Cargo.toml
@@ -8,4 +8,4 @@ edition = "2021"
 [dependencies]
 local_rules = { path = "../local_rules" }
 tree-sitter-lint = { path = "../.." }
-tree-sitter-lint-plugin-replace-foo-with-bar = { path = "../../tests/fixtures/tree-sitter-lint-plugin-replace-foo-with-bar" }
+tree-sitter-lint-plugin-replace-foo-with = { path = "../../tests/fixtures/tree-sitter-lint-plugin-replace-foo-with" }
