--- tests/fixtures/tree-sitter-lint-plugin-replace-foo-with-bar/src/lib.rs
@@ -1,32 +1,0 @@
-use std::sync::Arc;
-
-use tree_sitter_lint::{rule, violation, Plugin, Rule};
-
-pub fn instantiate() -> Plugin {
-    Plugin {
-        name: "replace-foo-with-bar".to_owned(),
-        rules: vec![replace_foo_with_bar_rule()],
-    }
-}
-
-fn replace_foo_with_bar_rule() -> Arc<dyn Rule> {
-    rule! {
-        name => "replace-foo-with-bar",
-        fixable => true,
-        listeners => [
-            r#"(
-              (identifier) @c (#eq? @c "foo")
-            )"# => |node, context| {
-                context.report(
-                    violation! {
-                        node => node,
-                        message => r#"Use 'bar' instead of 'foo'"#,
-                        fix => |fixer| {
-                            fixer.replace_text(node, "bar");
-                        }
-                    }
-                );
-            }
-        ]
-    }
-}
