+++ tests/fixtures/tree-sitter-lint-plugin-replace-foo-with/.gitignore
